Implementando serviços REST de cidades e estados

+++ src/main/java/com/jlcb/fordelivfood/infrastructure/repository/CidadeRepositoryImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,27 +17,32 @@ public class CidadeRepositoryImpl implements CidadeRepository {
 
 	@PersistenceContext
 	private EntityManager entityManager;
-	
+
 	@Override
 	public List<Cidade> listar() {
 		return entityManager.createQuery("FROM Cidade", Cidade.class).getResultList();
 	}
-	
+
 	@Override
 	public Cidade buscar(Long id) {
 		return entityManager.find(Cidade.class, id);
 	}
-	
+
 	@Override
 	@Transactional
 	public Cidade salvar(Cidade cidade) {
 		return entityManager.merge(cidade);
 	}
-	
-	@Override
+
 	@Transactional
-	public void remover(Cidade cidade) {
-		cidade = buscar(cidade.getId());
+	@Override
+	public void remover(Long id) {
+		Cidade cidade = buscar(id);
+
+		if (cidade == null) {
+			throw new EmptyResultDataAccessException(1);
+		}
+
 		entityManager.remove(cidade);
 	}
